App 컴포넌트화, context 모듈화
